Use corePoolSize = maxPoolSize and unbounded queue
Remove hack with calling setCorePoolSize to spawn new threads when already
corePoolSize is running. For bigger sizes of the ThreadPool setCorePoolSize took
a long time and blocked EDT thread. Now we will have always `maximum concurren
downloads` threads running for TMS layers, but this is not a big issue, as the
thread pool is shared between imagery layers.

As long as number of threads will not reach the `maximum concurrent downloads`
the `maximum concurrent downloads per host` limit will not be enforced (first n
downloads).

See: #16734



git-svn-id: https://josm.openstreetmap.de/svn/trunk@14261 0c6e7542-c601-0410-84e7-c038aed88b3b