see #17055 Disable autofix for possibly misspelled tag values
Some keys like shop have a long list of documented values and not all are in the presets. It can happen that a good value is flagged and a fix and an improper fix value is suggested. 
This should be reverted once we have a better source for "correct" values. The presets are too incomplete.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@14585 0c6e7542-c601-0410-84e7-c038aed88b3b