fix #17354: Relation list is not always updated
The problem was that `RelationListDialog.otherDatasetChange(AbstractDatasetChangedEvent event)` did nothing.
The dialog therefore misses events of type PRIMITIVE_FLAGS_CHANGED and if the flags of a relation are changed this should change the list.
When the dialog is build from scratch it looks at the relations with status complete and shows them.



git-svn-id: https://josm.openstreetmap.de/svn/trunk@14886 0c6e7542-c601-0410-84e7-c038aed88b3b